Patch by jnut and tbenita.
svn path=/plone.transforms/trunk/; revision=16912

+++ plone/transforms/command.py
@@ -1,5 +1,6 @@
 from logging import DEBUG
 import os
+from os.path import isdir
 import shutil
 import tempfile
 from cStringIO import StringIO as cStringIO
@@ -131,8 +132,8 @@ class CommandTransform(PersistentTransform):
                 fd.close()
                 os.unlink(tmp)
         finally:
-            os.unlink(tmpfilepath)
-            shutil.rmtree(tmpdirpath)
+            if isdir(tmpdirpath):
+                shutil.rmtree(tmpdirpath)
 
         return result
